Move APS status fields into TABLES objects, rename formConfig to fieldConfig, add helper to uppercase 'Aps' and call in formatDataTitles helper in Assets.js
